Create a mcbette report, to be used when reporting bugs

Create a mcbette report, to be used when reporting bugs

mcbette_report(beast2_folder = beastier::get_default_beast2_folder())

Arguments

beast2_folder
the folder where the BEAST2 is installed. Note that this is not the folder where the BEAST2 executable is installed: the BEAST2 executable is in a subfolder. Use get_default_beast2_folder to get the default BEAST2 folder. Use get_default_beast2_bin_path to get the full path to the default BEAST2 executable. Use get_default_beast2_jar_path to get the full path to the default BEAST2 jar file.

Value

nothing. It is intended that the output (not the return value) is copy-pasted from screen.

Author

Richèl J.C. Bilderbeek

Examples

mcbette_report()

#> ***********

#> * mcbette *

#> ***********

#> Can run mcbette: FALSE

#> OS: unix

#> ****************

#> * Dependencies *

#> ****************

#> beautier version: 2.6.1

#> tracerer version: 2.2.2

#> beastier version: 2.4.3

#> mauricer version: 2.5.1

#> babette version: 2.3

#> **********

#> * BEAST2 *

#> **********

#> Java version: 11.0.11+9-Ubuntu-0ubuntu2.20.04

#> Is BEAST2 installed: FALSE

#> ****************

#> * session info *

#> ****************

#> Warning: running command 'timedatectl' had status 1

#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value                       
#>  version  R version 4.1.0 (2021-05-18)
#>  os       Ubuntu 20.04.2 LTS          
#>  system   x86_64, linux-gnu           
#>  ui       X11                         
#>  language (EN)                        
#>  collate  en_US.UTF-8                 
#>  ctype    en_US.UTF-8                 
#>  tz       /UTC                        
#>  date     2021-07-15                  
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version    date       lib
#>  beastier      2.4.3      2021-07-15 [1]
#>  beautier      2.6.1      2021-07-15 [1]
#>  cachem        1.0.5      2021-05-15 [2]
#>  callr         3.7.0      2021-04-20 [2]
#>  cli           3.0.0      2021-06-30 [1]
#>  crayon        1.4.1      2021-02-08 [2]
#>  curl          4.3.2      2021-06-23 [1]
#>  desc          1.3.0      2021-03-05 [2]
#>  devtools      2.4.2      2021-06-07 [1]
#>  digest        0.6.27     2020-10-24 [2]
#>  downlit       0.2.1      2020-11-04 [2]
#>  ellipsis      0.3.2      2021-04-29 [2]
#>  evaluate      0.14       2019-05-28 [2]
#>  fansi         0.5.0      2021-05-25 [1]
#>  fastmap       1.1.0      2021-01-25 [2]
#>  fs            1.5.0      2020-07-31 [2]
#>  glue          1.4.2      2020-08-27 [2]
#>  htmltools     0.5.1.1    2021-01-22 [2]
#>  httr          1.4.2      2020-07-20 [2]
#>  knitr         1.33       2021-04-24 [2]
#>  lifecycle     1.0.0      2021-02-15 [2]
#>  magrittr      2.0.1      2020-11-17 [2]
#>  mauricer      2.5.1      2021-07-15 [1]
#>  mcbette     * 1.14       2021-07-15 [1]
#>  memoise       2.0.0      2021-01-26 [2]
#>  pillar        1.6.1      2021-05-16 [2]
#>  pkgbuild      1.2.0      2020-12-15 [2]
#>  pkgconfig     2.0.3      2019-09-22 [2]
#>  pkgdown       1.6.1      2020-09-12 [2]
#>  pkgload       1.2.1      2021-04-06 [2]
#>  prettyunits   1.1.1      2020-01-24 [2]
#>  processx      3.5.2      2021-04-30 [2]
#>  ps            1.6.0      2021-02-28 [2]
#>  purrr         0.3.4      2020-04-17 [2]
#>  R6            2.5.0      2020-10-28 [2]
#>  ragg          1.1.2      2021-03-17 [2]
#>  rappdirs      0.3.3      2021-01-31 [2]
#>  remotes       2.4.0      2021-06-02 [1]
#>  rJava         1.0-4      2021-04-29 [1]
#>  rlang         0.4.11     2021-04-30 [2]
#>  rmarkdown     2.9        2021-06-15 [1]
#>  rotemplate    0.0.1.9000 2021-05-20 [2]
#>  rprojroot     2.0.2      2020-11-15 [2]
#>  rstudioapi    0.13       2020-11-12 [2]
#>  sessioninfo   1.1.1      2018-11-05 [2]
#>  systemfonts   1.0.2      2021-05-11 [2]
#>  testit        0.13       2021-04-14 [1]
#>  testthat      3.0.4      2021-07-01 [1]
#>  textshaping   0.3.4      2021-05-11 [2]
#>  tibble        3.1.2      2021-05-16 [2]
#>  usethis       2.0.1      2021-02-10 [2]
#>  utf8          1.2.1      2021-03-12 [2]
#>  vctrs         0.3.8      2021-04-29 [2]
#>  whisker       0.4        2019-08-28 [2]
#>  withr         2.4.2      2021-04-18 [2]
#>  xfun          0.24       2021-06-15 [1]
#>  xml2          1.3.2      2020-04-23 [2]
#>  yaml          2.2.1      2020-02-01 [2]
#>  source                                   
#>  https://ropensci.r-universe.dev (R 4.1.0)
#>  https://ropensci.r-universe.dev (R 4.1.0)
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  https://ropensci.r-universe.dev (R 4.1.0)
#>  local                                    
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  Github (ropensci/rotemplate@24f57e6)     
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  RSPM (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#>  CRAN (R 4.1.0)                           
#> 
#> [1] /cache
#> [2] /usr/local/lib/R/site-library
#> [3] /usr/lib/R/site-library
#> [4] /usr/lib/R/library