Set the mcbette state.

Set the mcbette state to having BEAST2 installed with or without installing the BEAST2 NS package.

set_mcbette_state(
  mcbette_state,
  beast2_folder = beastier::get_default_beast2_folder(),
  verbose = FALSE
)

Arguments

mcbette_state
the mcbette state, which is a list with the following elements:

beast2_installed TRUE if BEAST2 is installed, FALSE otherwise

ns_installed NA if BEAST2 is not installed. TRUE if the BEAST2 NS package is installed FALSE if the BEAST2 NS package is not installed

beast2_folder
the folder where the BEAST2 is installed. Note that this is not the folder where the BEAST2 executable is installed: the BEAST2 executable is in a subfolder. Use get_default_beast2_folder to get the default BEAST2 folder. Use get_default_beast2_bin_path to get the full path to the default BEAST2 executable. Use get_default_beast2_jar_path to get the full path to the default BEAST2 jar file.

verbose
if TRUE show debug output

Note

In newer versions of BEAST2, BEAST2 comes pre-installed with the BEAST2 NS package. For such a version, one cannot install BEAST2 without NS. A warning will be issues if one intends to only install BEAST2 (i.e. without the BEAST2 NS package) and gets the BEAST2 NS package installed as a side effect as well.

Also, installing or uninstalling a BEAST2 package from a BEAST2 installation will affect all installations.

See also

Examples

mcbette_state <- get_mcbette_state()
mcbette_state$beast2_installed <- TRUE
mcbette_state$ns_installed <- TRUE
# \donttest{
  set_mcbette_state(mcbette_state)

#> Error in loadNamespace(x): there is no package called ‘beastierinstall’

# }